+++ src/lib.rs
@@ -1,3 +1,18 @@
+mod authorizer;
+mod bundler;
+mod exporter;
+mod importer;
+mod initializer;
 mod kibana_object_manager;
+mod manifest;
+mod merger;
+mod objects;
 
+pub use authorizer::Authorizer;
+pub use bundler::Bundler;
+pub use exporter::Exporter;
+pub use importer::Importer;
+pub use initializer::Initializer;
 pub use kibana_object_manager::KibanaObjectManagerBuilder;
+pub use manifest::Manifest;
+pub use merger::Merger;
