+++ src/manifest.rs
